up GameDb set full data in. -create a unique game between 2 players. -set time when was created. -add a column isProgress ,which tell me when a game is in still progress.

+++ demo-battleship/src/main/java/com/example/demo_battleship/repository/MoveRepository.java
@@ -3,5 +3,5 @@ package com.example.demo_battleship.repository;
 import com.example.demo_battleship.model.Move;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MoveRepository extends JpaRepository<Move, Long> {
+public interface MoveRepository extends JpaRepository<Move, Integer> {
 }
